모든 scene에 레이아웃 적용
